refactor: Name을 participant 패키지로 이동
