+++ src/main/java/domain/participant/Dealer.java
@@ -1,7 +1,5 @@
 package domain.participant;
 
-import domain.Name;
-
 public class Dealer extends Participant {
 
     private static final String DEALER_NAME = "딜러";
