+++ src/test/java/domain/participant/ParticipantTest.java
@@ -2,7 +2,6 @@ package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.Name;
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Symbol;
