+++ src/test/java/domain/participant/PlayerTest.java
@@ -2,7 +2,6 @@ package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import domain.Name;
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Symbol;
